pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "llm-wiki"
version = "0.1.0"
description = "AI-powered personal knowledge base built on LLM-maintained Markdown wiki"
requires-python = ">=3.10"
dependencies = [
    "click>=8.1",
    "openai>=1.0",
    "pyyaml>=6.0",
    "fastapi>=0.110",
    "python-multipart>=0.0.9",
    "uvicorn[standard]>=0.29",
]

[project.scripts]
llm-wiki = "llm_wiki.cli:main"

[tool.setuptools.packages.find]
include = ["llm_wiki*"]
